refactor(services): move process fetch helpers into a separated file
