+++ src/services/fetch.ts
@@ -1,0 +1,15 @@
+import { Observable } from 'rxjs/internal/Observable'
+import { throwError } from 'rxjs/internal/observable/throwError'
+
+export const processFetchResponse = (
+  response: Response
+): Promise<any> | Observable<Error> => {
+  if (response.ok) {
+    return response.json()
+  }
+
+  return throwError(new Error(`Error ${response.status}`))
+}
+
+export const processFetchError = (err: Error): Observable<Error> =>
+  throwError(new Error(err.message))
