+++ src/services/venues.ts
@@ -1,26 +1,12 @@
 import { StateObservable } from 'redux-observable'
 import { fromFetch } from 'rxjs/fetch'
-import { Observable } from 'rxjs/internal/Observable'
-import { throwError } from 'rxjs/internal/observable/throwError'
 import { catchError, switchMap } from 'rxjs/operators'
 import { NAction } from '../../types/action.d'
 import { NStore } from '../../types/store'
 import { EApiDefaultParameters, EApiPathnames } from '../constants/api'
 import { credentialsSelector } from '../selectors/life'
 import { getLocationHref } from '../utils/url'
-
-export const processFetchResponse = (
-  response: Response
-): Promise<any> | Observable<Error> => {
-  if (response.ok) {
-    return response.json()
-  }
-
-  return throwError(new Error(`Error ${response.status}`))
-}
-
-export const processFetchError = (err: Error): Observable<Error> =>
-  throwError(new Error(err.message))
+import { processFetchError, processFetchResponse } from './fetch'
 
 export const getObservableVenuesSearch = (
   action: NAction.IAction,
